chore(react): rename i18n provider module
