+++ packages/react/src/i18n/index.ts
@@ -26,4 +26,4 @@ export type {
 
 export { createI18n } from './create-i18n';
 export { I18nContext, useLocale, useTranslator } from './instance';
-export { I18nProvider } from './provider-instance';
+export { I18nProvider } from './provider';
